CellAutoFM/pyproject.toml
[project]
name = "cellautofm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.6.0",
    "evaluate>=0.4.4",
    "peft==0.13.2",
    "pillow>=11.3.0",
    "scikit-learn>=1.7.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "transformers>=4.53.0",
]
